+++ tools/blenderhelper.py
