+++ ydr/ydrimport.py
@@ -6,9 +6,9 @@ from Sollumz.ydr.shader_materials import create_shader
 from Sollumz.ybn.ybnimport import composite_to_obj
 from Sollumz.sollumz_properties import SOLLUMZ_UI_NAMES, BoundType, DrawableType, LODLevel, TextureFormat, TextureUsage
 from Sollumz.resources.drawable import *
-from Sollumz.meshhelper import flip_uv
+from Sollumz.tools.meshhelper import flip_uv
 from Sollumz.tools.utils import ListHelper
-from Sollumz.tools.blender_helper import BlenderHelper
+from Sollumz.tools.blenderhelper import BlenderHelper
 
 
 def shadergroup_to_materials(shadergroup, filepath):
